+++ modules/sor/lib/poolsV3/gyroECLP/gyroECLPPool.ts
@@ -35,10 +35,10 @@ export class GyroECLPPool extends BasePoolV3 implements BasePoolMethodsV3 {
     ): GyroECLPPool {
         const poolTokens: GyroPoolToken[] = [];
 
-        if (!pool.dynamicData) throw new Error('GyroECLP pool has no dynamic data');
+        if (!pool.dynamicData) throw new Error(`${pool.type} pool has no dynamic data`);
 
         for (const poolToken of pool.tokens) {
-            if (!poolToken.priceRate) throw new Error('GyroECLP pool token does not have a price rate');
+            if (!poolToken.priceRate) throw new Error(`${pool.type} pool token does not have a price rate`);
             const token = new Token(
                 parseFloat(chainToIdMap[pool.chain]),
                 poolToken.address as Address,