+++ modules/sor/lib/poolsV3/reClamm/types.ts
@@ -1,0 +1,11 @@
+export type ReClammParams = {
+    lastTimestamp: bigint;
+    lastVirtualBalances: bigint[];
+    dailyPriceShiftBase: bigint;
+    centerednessMargin: bigint;
+    startFourthRootPriceRatio: bigint;
+    endFourthRootPriceRatio: bigint;
+    priceRatioUpdateStartTime: bigint;
+    priceRatioUpdateEndTime: bigint;
+    currentTimestamp: bigint;
+};