+++ tasks/index.ts
@@ -15,6 +15,7 @@ import {
     TokenController,
     QuantAmmController,
 } from '../modules/controllers';
+import { PoolAprUpdaterService } from '../modules/pool/lib/pool-apr-updater.service';
 import { chainIdToChain } from '../modules/network/chain-id-to-chain';
 
 import { backsyncSwaps } from './subgraph-syncing/backsync-swaps';
@@ -187,6 +188,22 @@ async function run(job: string = process.argv[2], chainId: string = process.argv
         initRequestScopedContext();
         setRequestScopedContextValue('chainId', chainId);
         return poolService.reloadAllPoolAprs(chain);
+    } else if (job === 'update-pool-aprs') {
+        const id = process.argv[4];
+        const chain = chainIdToChain[chainId];
+        const service = new PoolAprUpdaterService();
+        const pools = await prisma.prismaPool.findMany({
+            where: { id: id, chain: chain },
+            include: {
+                dynamicData: true,
+                tokens: {
+                    include: {
+                        token: true,
+                    },
+                },
+            },
+        });
+        return service.updateAprsForPools(pools);
     } else if (job === 'update-prices') {
         await tokenService.syncTokenContentData(chain);
         return tokenService.updateTokenPrices([chain]);